Don't check test coverage in CI
While I'd still like some kind of process to keep an eye on test
coverage, the current status isn't really useful. It's slowing down our
CI while we suspect that nobody really checks it in the CI. It also
creates a big block of "noise" in the middle of pytest's report of
failed tests.

Locally, I see a speed difference with and without coverage
(with 4m6s vs without 3m14s).

Suggested & discussed in the last community meeting (2024-10-22).