Update _blur_effect.py
Apologies. I hope I got it right this time.  This is all new for me.  I certainly see the reasoning for moving it up a line. Why waste time doing a calculation if the value of M1 is zero.  As for the rest of the process in committing a change and other processes involved please let me know what I need to do.  As I said this is all new to me.