Explicitly set deprecation warning stacklevel and test
Apparently, `deprecate_parameter` has problems automatically detecting
the appropriate stacklevel when used in Cython code. Luckily, the
`stacklevel` parameter exists and allows us to sidestep the issue by
explicitly setting it. Also add a test, covering the deprecation
warnings. We only need to test the warnings and not the behavior of
the function, since the parameters previously did nothing.