Make apigen.py work with editable hooks
apigen.py in the previous form wasn't able to deal with skimage if it
was installed in editable mode using a _scikit-image_editable_loader
hook. In that case `root_module.__path__[-1]` would point to an invalid
directory.

This fix aims to (a) account for this by falling back to the __file__
attribute and (b) adds some defensive checks so we are not suprised in
the future when something breaks.

Longterm this part of our tooling could probably use some love, it seems
a bit brittle.