Test docs & gallery on Windows (Azure) only for `main`
The Azure jobs testing our documentation are still the longest running
jobs in our CI: ~40 min or longer. As a first step, I'd like to exclude
them from running for every PR. That should get the iteration time for
PRs down to 20 min.

@stefanv, suggested to disable them altogether. I'm not sure yet how I
feel about that. Since we do still tell contributors how to build the
docs on Windows, I do want to ensure that that keeps working. So for
now, I propose this compromise which keeps building the docs on `main`.