Ensure pytest configuration in pyproject.toml is picked up
Stepping out of the directory during tests might have been for the
purpose of ensuring that the installed version is tested? Not sure if
that is still necessary with the way we build and test. There's only
one compiled version so if pytest tries to do the wrong thing, then
it should fail and we should notice...

Though I would feel more secure about this if we were using the
SRC-layout.