Explicitly upcast float16 to float32 in random_walker (#7655)
* Explicitly upcast float16 to float32 with warning in random_walker

SciPy sparse, which is used in `random_walker`, doesn't officially
support float16. This led to failures when testing with NumPy 1.26, and
only worked accidentally for other NumPy versions (see GH issue 7635).

This fix handles that situation gracefully and makes the upcasting
clear to users, in the hopefully least disruptive way.

Also update related warning plubming from `expected_warnings` to
external `pytest.mark.filterwarnings`. The latter is hopefully a step
towards less maintenance.

* Ignore bycatch warning on `linux-cp3.11-mini-req*`

* Also handle dtype for spacing in `random_walker`

Make sure it's always using the same dtype as `data`.

* Fix logic error in spacing validation

* Add missing colon to filterwarnings decorator

* Address SparseEfficiencyWarning

The warning in question:

SparseEfficiencyWarning: Changing the sparsity structure of a
csr_matrix is expensive. lil_matrix is more efficient.

Wrapping the created csr_array in lil_array makes the warning go away.
Though not sure if this is actually the right approach.

* Revert "Address SparseEfficiencyWarning"

This reverts commit 3f53733454e6e02c85ebbeb0819bccf8e6256f60.

* Don't check stacklevel of ignored warning

* Remove warning about upcasting to float32

We decided that trying to preserve float32 isn't worth the effort; there
wasn't really a guarantee that we did so previously. We just make sure
to upcast from float16 which might lead to failures.

* Remove obsolete warningfilters

This warning was removed in a previous commit.

* Remove obsolete filterwarnings (II)

This warning was removed in a previous commit.