Use pytest config in pyproject.toml in CI (#7555)
* Use pytest config in pyroject.toml in CI

This should be safe, as the pytest documentation [1] states that
"rootdir is NOT used to modify sys.path/PYTHONPATH". So while the
`--config-file` option affects the `rootdir`, it's not adding that
directory to the PYTHONPATH where it could accidentally overwrite the
installed package. This solution is somewhat unsatisfying. Personally,
I think a better approach would be to switch to the src-layout.

[1] https://docs.pytest.org/en/7.1.x/reference/customize.html?highlight=rootdir#initialization-determining-rootdir-and-configfile

* Use correct operator to check if GITHUB_WORKSPACE exists

* Ignore missing dask dependency in cycle_spin

* Don't expect newbyteorder warning caused by tifffile

I believe this warning disappeared with tifffile 2024.4.24 which added
support for NumPy 2.

* Ignore warning if cg_mg is not available (missing pyamg)

* Update test scripts in new locations

* Pin higher version of pytest

* Pin newer dask to get rid of distutils warning

* Expect sparse efficiency warnings until new pyamg

* New dask version requires newer cloudpickle

* Tell pytest where to find SparseEfficiencyWarning

* Catch additional warnings

* Up minimum version of SciPy to fix sparse errors

* Catch additional warning

* Ignore warnings raised by pyamg internally

* Use a reasonable range in simpleitk roundtrip test

* Re-enable testing in Emscripten/Pyodide job

Leaving this commented is probably an oversight from gh-7576 /
a7b54a87b77d9ef1558e8bb7d70992f4e0b69ab1.

---------

Co-authored-by: Stefan van der Walt <stefanv@berkeley.edu>