Revert `ball` to deprecated algorithm
If we keep this deprecated but around for a while it's probably a good
idea to revert to the original algorithm. I intend to cover the
backwards-compability of the new algo `footprint_ellipse` in a full test
but if there are edge cases we miss, this should keep users who stay on
the old version insulated.