Mark `disk`, `ball`, `ellipse` as deprecated
Unfortunately, there seem to be slight differences when trying to
reproduce `disk` and `ball` shaped footprints with `footprint_ellipse`.
Right now I'm guessing that this might be due to floating point
calculations.

`disk` and `ball` construct the equation in a different way, than
`footprint_ellipse` which supports N dimensions.

Pushing this now to facilitate discussion.