Improve mcp find costs (#7625)
* Add check for max_cost in MCP::find_costs

* Add description for max_cost of MCP::find_costs

* Add unit test for find_costs max_cost param

* Deprecate find_costs max_cumulative_cost param

* Rename `max_step` to `max_step_cost`

* keyword only parameters for find_costs

* Deprecate max_cost paramter for find_costs

* Simplify check

* Bump deprecations to 0.26.0

after 0.25.0 release. Also point to `new_name` for deprecated `max_cost`
parameter.

* Explicitly set deprecation warning stacklevel and test

Apparently, `deprecate_parameter` has problems automatically detecting
the appropriate stacklevel when used in Cython code. Luckily, the
`stacklevel` parameter exists and allows us to sidestep the issue by
explicitly setting it. Also add a test, covering the deprecation
warnings. We only need to test the warnings and not the behavior of
the function, since the parameters previously did nothing.

* Only make new parameter `max_step_cost` keyword-only

we need to keep previous parameters accessible by position in order to
not break users.

* Add wording suggestion by Jarrod

Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>

* Update TODO.txt

---------

Co-authored-by: Lars Grüter <lagru@mailbox.org>
Co-authored-by: Stefan van der Walt <stefan@mentat.za.net>
Co-authored-by: Jarrod Millman <jarrod.millman@gmail.com>