Deprecate `area_threshold` in `remove_small_holes`
In 0.16, in remove_small_holes, min_size has been replaced definitely
by area_threshold. But for remove_small_objects, it is still min_size.
This attempts to solve that inconsistency.

Having dug through the past issues and discussions on this, it seems to
me that each discussed option will be a trade-off between consistency
and matching the context of each function.

Since remove_small_objects and remove_small_holes are so similar, I
definitely give consistency more weight. Ignoring past deprecations,
min_size seems to be the least controversial option. The main argument
against it was that we'd have to revert a previous deprecation.

That's not ideal, but addressing previous mistakes is precisely what
skimage2 is about. So, I propose to deprecate area_threshold in
remove_small_holes back to min_size and only finalize the deprecation
in 2.0.0.

I think we can keep using area_threshold in area_opening and
area_closing, since area is a prominent term and I anticipate less user
confusion.