Make footprint cross decomposition a public function
Since this function is now exposed to all kinds of footprints, I tried
to add sanity checks that didn't impact the performance to much and
tried to come up with a sensible behavior for edge cases.

It also removes the `dtype` parameter. Instead it just uses the dtype
of `footprint`, allowing users to control the outputs dtype that way.
I couldn't really think of use cases where the parameter would be
more useful.