Use `max_size` in `remove_small_(holes|objects)` (#7739)
* Deprecate `area_threshold` in `remove_small_holes`

In 0.16, in remove_small_holes, parameter min_size was replaced
by area_threshold. But for remove_small_objects, it was still min_size.
This PR attempts to fix that inconsistency.

Having dug through the past issues and discussions on this, it seems to
me that each discussed option is be a trade-off between consistency
and matching the context of each function.

Since remove_small_objects and remove_small_holes are so similar, I
definitely give consistency more weight. Ignoring past deprecations,
min_size seems to be the least controversial option. The main argument
against it was that we'd have to revert a previous deprecation.

That's not ideal, but addressing previous mistakes is precisely what
skimage2 is about. So, I propose to deprecate area_threshold in
remove_small_holes back to min_size and only finalize the deprecation
in 2.0.0.

I think we can keep using area_threshold in area_opening and
area_closing, since area is a prominent term and I anticipate less user
confusion.

* Use min_size in tests instead of area_threshold

* Refactor to `max_size` in `remove_small_(objects|holes)`

> From remove_small_holes(a, min_size=...), I don't think a user can
> figure out what min_size means, whereas
> remove_small_holes(a, max_size=...) reads pretty clearly as
> "remove small holes, but don't go beyond max_size".

– Stéfan

* Use "contiguous" in docstring

* Make new `max_size` parameter keyword-only;
as a side effect, the existing argument "connectivity" can be made
keyword-only as well.

* Improve parameter description

---------

Co-authored-by: Stefan van der Walt <github@mentat.za.net>
Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>
Co-authored-by: Stefan van der Walt <stefanv@berkeley.edu>