Make stacklevel detection work for decorated methods too
and add `assert_stacklevel` checks to `test_fit.py`.

The first item makes `count_global_wrappers`, formerly `_get_stack_length`,
account for `func` being a method. This is necessary because we are now using
`deprecate_func` on the `estimate` method of Transform classes and such.

Also remove `_DecoratorBaseClass` which didn't add anything of value
and was just passing through values.