Fix _warning_stacklevel
There was some code duplication with `count_global_wrappers`. And I
suspect that the logic in the previous version was wrong:
```python
for name in func.__qualname__.split("."):
    global_func = getattr(module, name)
```
calling `getattr` on `module` repeatetly is wrong. Instead, consolidate
with `count_global_wrappers` and make sure that that one fails
transparently for closures. In that case, the stacklevel should be
set manually instead!