Move `_cross_entropy` to new _thresholding_global
This private function is used in the plot_threshold_li gallery example.
It's a bit weird but Stéfan suggests to keep it private for now since
the API doesn't look really useful. Instead he suggests that a better
API would be to compute the cross entropy between two arrays.
I concur and will open an issue.