Solve import cycle between segmentation & filters modules
Moving the implementation to `skimage.segmentation` and keeping the
deprecated API around with wrappers, introduced a dependency cycle with
`skimage.filters` now depending on `segmentation` which already
depends on `filters`. Lazy import inside the deprecated functions seems
to address this.

Note, this import cycle is only triggered if lazy_loader is set to
eager importing (env variable `EAGER_IMPORT=1`).

In the process also update where the default for
`_cross_entropy(..., bins=XXX)` is created. Using `None` is the more
common pattern. It also avoids the default for `bins` blowing up the
representation of the signature with a large tuple.