Back of to direct method deprecation of estimate
I was using some brittle docstring parsing of the `from_estimate` method
to make the deprecated `estimate` method, but now we need to document
the return type of `from_estimate` in examples, this won't be practical.
Copy back all the original `estimate` docstrings, and use these where
possible.  Where `estimate` is inherited from a base class, make a new
`estimate` method using the signature and docstring of the base class,
but where the deprecation message shows the correct class in suggestion
to use `from_estimate`.