Make sure import-mode is set only for doctests
--import-mode="importlib" is necessary to successfully collect doctests.
Otherwise, the doctest-plus plugin will fail because it finds doctests
in skimage/ which don't exist at runtime. E.g, `brief_pythran.py` which
exists as a file but is only available at runtime with `brief_cy`.

However, there's no need to deviate from the default import mode if
doctests are not collected.