Use PendingDeprecationWarning approach for skimage2
The new approach uses the custom warning category
`PendingSkimage2Change`. As a subclass of Python's
`PendingDeprecationWarning` it is not shown by default. This means that
for now users can explicitly opt in to receive warnings related to
skimage2 and address them. At a later stage we can switch these to a
subclass of `FutureWarning` so they are shown by default.

I like this approach because for now, we are still figuring out exactly
how we will manage the transition process. By using
`PendingDeprecationWarning` we can still update that process without
annoying the average user unnecessarily. Once we have a clear picture,
we can simply switch to using `FutureWarning` and everything is in
place.

It also differentiates skimage2 related warnings from conventional
deprecation warnings. Something we really want because both categories
have different consequences.