Deprecate estimate method in favor of class constructor (#7771)
## Description

This is the promised follow-on PR from the discussion more or less
starting at
https://github.com/scikit-image/scikit-image/pull/7754#issuecomment-2742974392.

To recap briefly:

There is a wart in the current API, in that the `estimate` method of
Transforms ignores any stored state of the initialized Transform, and
re-estimates. So, for example:

```python
tf = SimilarityTransform(rotation=(0.2, 0.3))
tf.estimate(src, dst)
```

wil discard the initial state with the specified rotations. Another
problem is that one generally has to first initialize the transform,
most sensibly as an identity transform, and then estimate, thus:

```python
tf = SimilarityTransform.identity(3)
tf.estimate(src, dst)
```

However, the argument to the identity transform is redundant, because we
can infer the dimensionality of the points (3) from `src` and `dst`
point arrays. So, this PR deprecates the `estimate` method, in favor of
a class constructor, like this:

```python
tf = SimilarityTransform.from_estimate(src, pts)
```

### For discussion

* We floated various names for the class constructor. `from_estimate`
gained favor from me (MB) (I proposed it) and @stefanv. @jni leaned
towards some fancy footwork to allow `estimate` to be *both* the
original estimation method, *and* the class constructor, using the
[trick
here](https://stackoverflow.com/questions/20533349/python-regular-method-and-static-method-with-same-name/28766809#28766809).
I [thought that was going to be
confusing](https://github.com/scikit-image/scikit-image/pull/7754#issuecomment-2748623196).
We can rehearse that argument more.
* In writing the tests, I (MB) discovered an error in the check for a
degenerate tranform, [discussed
here](https://github.com/scikit-image/scikit-image/issues/7770). I used
an obvious but quite possibly wrong workaround (mentioned in the linked
issue). I would be happy to be corrected.
* One wrinkle is that the `from_estimate` API returns `None` for failed
estimation. This is generally just as informative as the `estimate`
method, except in the case of `PiecewiseAffineTransform`, where the
estimation involves multiple steps, and resulting affines. The
`estimate` method collects an affine for each step, recording an all-NaN
affine for steps that fail estimation, so you can track, from the
resulting `tform.affines` attribute, where the estimation failed. The
returned `None` does not give you that granular information. Is there a
use for that information? How should it reach the user?
* Should we also adapt the API-a-like classes `LineModelND`,
`CircleModel`, `EllipseModel` at
https://github.com/scikit-image/scikit-image/blob/main/skimage/measure/fit.py#L27
to have a `from_estimate` class method (instead of an `esimate` method)?
* Similarly, should we also adapt the `ThinPlateSpline` class
(https://github.com/scikit-image/scikit-image/blob/main/skimage/transform/_thin_plate_splines.py#L7)?
(I think so).

## Release note

For maintainers and optionally contributors, please refer to the
[instructions](https://scikit-image.org/docs/stable/development/contribute.html#documenting-changes)
on how to document this PR for the release notes.

```release-note
In `skimage.measure`, add a new class method and constructor `from_estimate` for 
`LineModelND`, `CircleModel`, and `EllipseModel`. This replaces the old more un-intuitive
API that used the now deprecated `estimate` method, which required initalizing a model
with undefined parameters before calling `estimate`. 
```

```release-note
In `skimage.transform`, add a new class method and constructor `from_estimate` for
`AffineTransform`, `EssentialMatrixTransform`, `EuclideanTransform`,
FundamentalMatrixTransform`, `PiecewiseAffineTransform`, `PolynomialTransform`, 
`ProjectiveTransform`, `SimilarityTransform`, and `ThinPlateSplineTransform`. This
replaces the old more un-intuitive API that used the now deprecated `estimate` method,
which required initalizing an undefined transform before calling `estimate`.
```

---------

Co-authored-by: Lars Grüter <lagru@mailbox.org>
Co-authored-by: Stefan van der Walt <stefan@mentat.za.net>