Setup stub creation with docstub in CI (#7802)
This purposefully doesn't address any of the errors yet. And instead
pins the number of errors to an upper bound. The plan is to address
these bit by bit in successive PRs and probably also by improving the
capabilities of docstub itself as we go.

After we've addressed all docstub errors, we can start up a type checker
and think about distributing the stubs. :blush:

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>
Co-authored-by: Stefan van der Walt <sjvdwalt@gmail.com>