Describe custom warning-strategy in migration guide
These changes are extracted from [1] since that PR will probably not be
merged. Nevertheless, the machinery and idea around using a custom
warning for skimage2 related changes is useful. Early on we can make
it subclass `PendingDeprecationWarning` which will not be shown by
default [2]. However, it can easily be made visible in case we or users
want to take an early look and prepare. Later, we can switch the base
out to `FutureWarning` to make it visible by default.

[1] https://www.github.com/scikit-image/scikit-image/pull/7797
[2] https://docs.python.org/3/library/exceptions.html#PendingDeprecationWarning

Co-authored-by: Marianne Corvellec <marianne.corvellec@ens-lyon.org>
Co-authored-by: Stefan van der Walt <stefanv@berkeley.edu>