Describe custom warning-strategy in migration guide (#7857)
These changes are extracted from
https://github.com/scikit-image/scikit-image/pull/7797 since that PR
will probably not be merged. Nevertheless, the machinery and idea around
using a custom warning for skimage2 related changes should be useful.
Early on we can make it subclass
[`PendingDeprecationWarning`](https://docs.python.org/3/library/exceptions.html#PendingDeprecationWarning)
which will not be shown by default. However, it can easily be made
visible in case we or users want to take an early look and prepare.
Later, we can switch the base out to `FutureWarning` to make it visible
by default.