Include "skimage" as filtered module again in warning filter
I'm not exactly sure why this is necessary to catch "skimage" too. Maybe
some part of the implementation raises the warning internally and
doesn't catch it. At the worst, this means users might see a warning
twice. I don't really care enough to investigate because we are
removing the whole thing.