Fix test to not fetch data already presented in test directory, fix tests on Python 3.14 and macos arm  (#7881)
Extracted from #7877 

Currently the tests in building wheels are failing because `pytest` is
not called from the root directory of the repository.

This PR contains fixes for Python 3.14 and loading test data without
fetch.

We're not quite sure why yet, but on macOS ARM some results differ from x86_64 after a few digits, so relaxed test precision.

---------

Co-authored-by: Lars Grüter <lagru@mailbox.org>