Add information about time of execution of test on each file (#7874)
Use code from napari to report the time of execution of each file during tests.
This allows us to easily trace which part of the test suite is slow.
The reporting is enabled by the CI environmental variable.

Also adds the pytest-pretty dependency to improve the readability of the test report.