Add dependabot configuration (#7882)
This PR adds dependabot [1] configuration to the scikit-image repository.

The dependabot generates automatic PRs to bump action versions. 

This reduces maintenance time, as maintainers do not need to check if
there is a new version of actions. It is configured to perform monthly
checks to not generate too many PRs.

You may see an example of PR in the napari repository:
https://www.github.com/napari/napari/pull/7572


[1] https://docs.github.com/en/code-security/dependabot/dependabot-security-updates/configuring-dependabot-security-updates
---------

Co-authored-by: Lars Grüter <lagru+github@mailbox.org>
Co-authored-by: Lars Grüter <lagru@mailbox.org>