Address ResourceWarning in _built_utils/version.py
The opened `__init__.py` was never properly closed. In certain cases
this caused pytest to fail with an ResourceWarning – because Pytest
would filter all warnings as errors.

In the process, clean up the script and move it to tools/ where it is
way more appropriate. Should also have way less potential for weird
interactions and confusion since it is no longer in our package.