Use pytest's summary over pytest-pretty's table-based one
Personally, I feel like the table format is a strict downgrade from the
default format that wraps and works with any terminal with. Even if we
try to increase the table width, it will still hide the full file path
and long test names.

I noticed that pytest-pretty subclasses pytest's TerminalReporter which
is marked as "@final". It seems like pytest-pretty is not making use of
the endorsed hook mechanism here. I hope that this won't be too brittle
going forward.

If it becomes a maintenance burden, we should feel free to rip it out
again instead of fixing it.