Extract `_minmax_scale_value_range` into its own function
`_prescale_value_range` is really more of a helper function that ties
our old legacy scaling behavior and the new min-max scaling together.
It is meant as a convenience function to use in a few places where
our API needs to work with a well-behaved output range. We very
probably don't want to the slightly awkward API of
`_prescale_value_range` public.

Instead, if we want to later make something public. It's probably
something closer to a fresh API like `_minmax_scale_value_range`.