Document and support `boundary_mode` 'nearest', 'wrap', 'mirror'
I found it a bit tricky to describe the boundary modes in terms of their
effects on the resulting blobs. It's tempting to delegate to the
docstring of `gaussian` but that isn't really helpful to users without
being aware of the actual algorithm or implementation.

I opted to not support all modes that gaussian supports. I'm not sure
how useful they would really be in addition to 'nearest' and 'wrap'.

If I could switch defaults, I would probably switch to 'wrap' or
'constant' over 'nearest'. 'nearest' has a significant effect but only
for blobs whose center lies exactly on the edge.