Test fixes and improvmement from cibuldwheel PR (#7922)
We are attempting to refactor our CI wheel building configuration 
which is based on cibuildwheel in gh-7922. This extracts a few 
tests fixes and tweaks that are necessary for gh-7922 to pass.

Especially, switching from `expected_warnings` to `pytest.warns` avoids
triggering a deprecation warning in `numpy.linalg` because the
implementation of `expected_warnings` accesses `__warningregistry__`.