Allow Cython 3.2.0 again [skip ci]
With the fix for https://www.github.com/cython/cython/issues/7259 was
merged into Cython's master. We should eventually be able to run our
test suite with Cython 3.2 again. It looks like the fix will be included
in 3.2.0r1 [1], so leave incompatibility with 3.2.0b1 in.

It looks like the nightly wheels repo [2] doesn't include the fix yet,
so skip the CI for now but create this PR as a reminder to revisit.

[1] https://github.com/cython/cython/commit/4e031392ba6e87203f8ee1f54e7c3ad336fd45d6
[2] https://pypi.anaconda.org/scientific-python-nightly-wheels/simple/