Relax constraints of regionprops multichannel test on MacOS with NumPy & "Accelerate" (#7942)
We were getting failures to given tolerance testing difference between
single-channel `moments_weighted_central`. This was because the
multi-channel array was not-contiguous, and `moments_central` used
`np.dot`, which in turns calls out the BLAS implementation. The BLAS
implementation result can depend on array memory ordering, and this was
the case for MacOS Intel Accelerate, in particular.

To reduce this variation between single-channel and multi-channel, use
`np.einsum` to replace `np.dot`; this also allows us to avoid the
`np.rollaxis` calls to align the summing axes for `np.dot`.

Closes #7941

---------

Co-authored-by: Lars Grüter <lagru+github@mailbox.org>