Use __doctest_requires__ instead of inline importorskip
With pytest-doctestplus v1.6.0 [1], doctests that are skipped via
__doctest_requires__ are now visible in PyTests output. This addresses
one of the concerns @lagru had in adopting these.

Removing the inlined importorskip from docstring examples makes them
a bit less confusing for people taking a peak. I think this was a
concern brought up by @stefanv.

A disadvantage of this approach, is that doctests will fail if they
are run without the --doctestplus flag and dependencies are not
available – for example when invoking doctests through PyCharm's UI.

[1] https://github.com/scientific-python/pytest-doctestplus/releases/tag/v1.6.0