Add algorithm to defer points causing ties
Some points cause ties in the accumulator.

Sometimes this is because the points are on two possible lines, and
sometimes it is because, for the selected points, small variations in
angles catch some part of the line, and the accumulator picks this up as
nearby thetas.

Try deferring points causing threshold ties to the end of the point
list.

This problem was likely the cause of adaptation of the the simplest
initial test, by restricting thetas.  Remove this adaptation, it does
not appear to be necessary after the tie check.