Don't add `--ignore` option implicitly
After local testing, it seems that it has no effect for out-of-tree
builds. I remember it working in the past but maybe I missed something.

Because I can't make it work (--ignore-glob also doesn't), I'm instead
opting to print a helpful warning in case `src/` is passed with an
out-of-tree build.