Don't trigger Release wheels workflow on pyproject.toml change
I think in practice this file is changed far to often for unrelated
reasons. Thus resulting in a significant waste of resources. We are
periodically building and testing wheels for our nightlies. So we should
still catch stuff early. And if necessary there's always the option to
use a "maintenance/*" branch.